catalog-react: Remove useOwnedEntities hook
