+++ .changeset/spotty-seals-press.md
@@ -1,0 +1,7 @@
+---
+'@backstage/plugin-catalog-react': minor
+---
+
+**BREAKING**: Removed `useOwnedEntities` and moved its usage internally to the scaffolder-backend where it's used.
+
+**BREAKING**: Removed `EntityTypeReturn` type which is now inlined.
