+++ .changeset/young-feet-flow.md
@@ -1,0 +1,5 @@
+---
+'@backstage/plugin-catalog-react': patch
+---
+
+Internalized usage of `useOwnedEntities` hook.
