+++ plugins/catalog-react/api-report.md
@@ -13,7 +13,6 @@ import { ComponentEntity } from '@backstage/catalog-model';
 import { ComponentProps } from 'react';
 import { CompoundEntityRef } from '@backstage/catalog-model';
 import { Entity } from '@backstage/catalog-model';
-import { GetEntitiesResponse } from '@backstage/catalog-client';
 import { IconButton } from '@material-ui/core';
 import { LinkProps } from '@backstage/core-components';
 import { Observable } from '@backstage/types';
@@ -388,15 +387,6 @@ export interface EntityTypePickerProps {
   initialFilter?: string;
 }
 
-// @public @deprecated (undocumented)
-export type EntityTypeReturn = {
-  loading: boolean;
-  error?: Error;
-  availableTypes: string[];
-  selectedTypes: string[];
-  setSelectedTypes: (types: string[]) => void;
-};
-
 // @public
 export const FavoriteEntity: (props: FavoriteEntityProps) => JSX.Element;
 
@@ -547,12 +537,6 @@ export function useEntityTypeFilter(): {
   availableTypes: string[];
   selectedTypes: string[];
   setSelectedTypes: (types: string[]) => void;
-};
-
-// @public @deprecated
-export function useOwnedEntities(allowedKinds?: string[]): {
-  loading: boolean;
-  ownedEntities: GetEntitiesResponse | undefined;
 };
 
 // @public @deprecated
